requirements-api.txt
# FastAPI 後端額外依賴
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.7
websockets>=12.0
google-cloud-bigquery>=3.11.0
db-dtypes>=1.1.1
